DependencyGraphBuilder: Use an appropriate key for nodeIndexMapping
According to the documentation of `DependencyReference.kt`, the tuple
`pkg, fragment` is a unique key. Use that key for clarity. Amongst
others the mutable "dependencies" property isn't part of the hashing
anymore which ensures that the hash value remains constant. It also
improves the asymptotic complexity of that hashing to O(1) because only
the attributes of the respective node need to be hashed, not the whole
connected component of the (directed) graph.

Signed-off-by: Frank Viernau <frank.viernau@here.com>