DependencyGraphBuilder: Remove all cycles when building the graph
The assumption that the dependency graph does not contain any cycles is
hard-coded in several places / components accessing the dependency
graph, see e.g. [1],[2],[3] and [4].

In the long run access to the dependency graph should only happen via
dependency navigator API and maybe also still via the old data model
returned by DependencyGraph.createScopes(). The new dependency (graph)
data format allows representing circles and it makes sense to do so in
order to reflect the reality as close as possible.

However, the code places (mentioned above) assuming the absence of
cycles run into endless loops or out of memory when a cycle is present.
This infact was recently observed when scanning several Gradle projects.
While the long term solution should probably be allowing cycles and
removing mentioned problematic assumptions, guarantee the absence of any
cycle to fix all issues with cycles. This should be reverted as soon as
the right solution is in place.

[1] https://github.com/oss-review-toolkit/ort/blob/master/model/src/main/kotlin/DependencyGraph.kt#L198-L215
[2] https://github.com/oss-review-toolkit/ort/blob/master/model/src/main/kotlin/DependencyGraph.kt#L229-L252
[3] https://github.com/oss-review-toolkit/ort/blob/master/model/src/main/kotlin/DependencyNavigator.kt#L67-L84
[4] https://github.com/oss-review-toolkit/ort/blob/master/model/src/main/kotlin/DependencyNavigator.kt#L232-L235

Signed-off-by: Frank Viernau <frank.viernau@here.com>